requirements.txt
streamlit
google-generativeai
pandas
python-dotenv
pillow
supabase
pdf2image
